Move ProxyResponse to its own class so it can be reused
